third_party/WebKit/LayoutTests/fast/forms/number/number-change-event-by-defaultvalue.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="log"></div>
<input type="number">
<script>
test(function() {
    var input = document.querySelector('input[type=number]');
    var didDispatchChange = false;
    input.onchange = function() {
        didDispatchChange = true;
    };
    input.defaultValue = '123';
    input.focus();
    input.blur();
    assert_false(didDispatchChange);
    assert_equals(input.value, '123');
}, 'Setting defaultValue should not dispatch \'change\' event on blur.');
</script>
</body>
